+++ gesturefx-sample/src/main/java/net/kurobako/gesturefx/sample/ArbitraryNodeSample.java
@@ -13,26 +13,22 @@ import javafx.scene.SubScene;
 import javafx.scene.control.Label;
 import javafx.scene.layout.Priority;
 import javafx.scene.layout.VBox;
-import javafx.scene.web.WebView;
 
-public class ArbitraryNodeSample implements Sample{
+public class ArbitraryNodeSample implements Sample {
 	@Override
 	public Node mkRoot() {
 		try {
-			Parent node  = FXMLLoader.load(getClass().getResource("/ComplexScene.fxml"));
+			Parent node = FXMLLoader.load(getClass().getResource("/ComplexScene.fxml"));
 			node.setPickOnBounds(true);
-			node.setMouseTransparent(true);
-			WebView webview = (WebView) node.lookup("webview");
-			if(webview != null){
-				webview.getEngine().load("http://purecss3.net/doraemon/doraemon_css3.html");
-			}
+			node.setMouseTransparent(false);
 
 			GesturePane pane = new GesturePane(new SubScene(node, 500, 500));
 			VBox.setVgrow(pane, Priority.ALWAYS);
 			Label description = new Label("Zoom and scroll on the SubScene below, " +
-					                              "observe that controls in JavaFX are vectors " +
-					                              "and that lighting effects are respected" +
-					                              "(different zoom alters light distance).");
+					"observe that controls in JavaFX are vectors " +
+					"and that lighting effects are respected" +
+					"(different zoom alters light distance). Gestures still work when the mouse " +
+					"event is consumed(e.g the bottom black bar)");
 			description.setWrapText(true);
 			description.setPadding(new Insets(16));
 			return new VBox(description, pane);